# test(cmd/prometheus): speed up test execution by t.Parallel() when possible