related to https://github.com/prometheus/prometheus/issues/15185

turned some loops into subtests to make use of t.Parallel()

On my machine:

Reducing exec time for `github.com/prometheus/prometheus/cmd/prometheus` by `~65%`: `~110s -> ~35s`

After rebasing + adding more `t.Parallel()`:

```
$ go test -race --count=1 ./cmd/prometheus/
ok  	github.com/prometheus/prometheus/cmd/prometheus	13.862s
```
on main (98dcd28b1afe9a0db08dc00de090208e892ba473)
```
$ go test -race --count=1 ./cmd/prometheus/
ok  	github.com/prometheus/prometheus/cmd/prometheus	99.341s
```
`~-85%`


In the CI moved from `github.com/prometheus/prometheus/cmd/prometheus	34.132s` to `ok  	github.com/prometheus/prometheus/cmd/prometheus	93.065s` (`~-60%`)

- [ ] create an issue for the other packages.

<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->