@SuperQ 2024-08-16
Now that Go 1.23 is out, we can adopt Go 1.22 features.