@machine424 2024-08-20
Opened https://github.com/prometheus/prometheus/pull/14697